requirements.txt
langchain==0.0.184
PyPDF2==3.0.1
python-dotenv==1.0.0
streamlit==1.18.1
openai==0.27.6
faiss-cpu==1.7.4
altair==4
tiktoken==0.4.0
boto3==1.18.63
#huggingface-hub==0.0.12
#sentence-transformers==2.2.2
#InstructorEmbedding==1.0.1




# uncomment to use huggingface llms
# I tried to use the latest version of huggingface-hub but i sometimes had issues while launching it
# So if you need a free use llm model you might wanna still try using i don't guarantee that it is working properly

# huggingface-hub==0.14.1

# uncomment to use instructor embeddings
# InstructorEmbedding==1.0.1
# sentence-transformers==2.2.2



# To install the required packages, run the following commands in the terminal:
# pip install streamlit pypdf2 langchain python-dotenv faiss-cpu openai huggingface_hub

# In case you are missing some packages for huggingFace :
# pip install -U langchain-community
# pip install InstructorEmbedding sentence_transformers


#Pour le cloud pip install boto3
